backend/requirements.txt
# Django and REST Framework
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==24.1

# Image processing
Pillow==11.3.0

# Environment and configuration
python-decouple==3.8

# Development and testing
pytest==8.0.0
pytest-django==4.8.0

# Production (optional)
gunicorn==21.2.0
whitenoise==6.6.0

# Database (optional - for production)
# psycopg2-binary==2.9.9  # PostgreSQL
# mysqlclient==2.2.1       # MySQL

# Caching (optional)
# redis==5.0.1
# django-redis==5.4.0

# File storage (optional - for production)
# boto3==1.34.0            # AWS S3
# django-storages==1.14.2

# Monitoring and logging (optional)
# sentry-sdk==2.0.0 